Apply package migration handling to the core integration tests as well, and centralize it all in a single test base
