+++ src/Umbraco.Test.Search.Examine.Integration/Tests/ContentTests/IndexService/InvariantSortableIndexTests.cs
@@ -1,7 +1,6 @@
 ﻿using Examine;
 using Examine.Search;
 using NUnit.Framework;
-using Umbraco.Cms.Core;
 using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Search.Provider.Examine.Helpers;
 using Umbraco.Cms.Tests.Common.Builders;
@@ -35,12 +34,6 @@ public class InvariantSortableIndexTests : IndexTestBase
             Assert.That(values.Skip(1).First(), Is.EqualTo("B Title"));
             Assert.That(values.Last(), Is.EqualTo("C Title"));
         });
-    }
-
-    [SetUp]
-    public async Task RunMigrations()
-    {
-        await WaitForPackageMigrationsAsync();
     }
 
     private async Task CreateTitleDocType()
